Add transport→protocol timing metrics enhancing end-to-end pipeline observability
Add new timing metrics to measure key stages of the message pipeline:

ws_receive_block_duration – WebSocket receive blocking time (network wait)
ws_process_message_duration – transport processing time per message
message_handoff_duration – latency from transport commit to protocol consumption
message_process_duration – protocol parsing and routing time

These metrics provide a clear breakdown of the pipeline: network → transport → queue handoff → protocol

Enabled under telemetry L3 with no overhead when disabled.

+++ include/lcr/metrics/atomic/stats/life_cycle.hpp
@@ -62,16 +62,16 @@ struct alignas(64) life_cycle {
         return 100.0 * static_cast<double>(total_active_time_ns_.load()) / static_cast<double>(total);
     }
 
-    inline double avg_cycle_ns() const noexcept {
+    inline uint64_t avg_cycle_ns() const noexcept {
         const uint64_t n = cycle_count_.load();
-        if (n == 0) return 0.0;
-        return total_cycle_time_ns_.load() / static_cast<double>(n);
+        if (n == 0) return 0;
+        return total_cycle_time_ns_.load() / n;
     }
 
-    inline double avg_active_ns() const noexcept {
+    inline uint64_t avg_active_ns() const noexcept {
         const uint64_t n = cycle_count_.load();
-        if (n == 0) return 0.0;
-        return total_active_time_ns_.load() / static_cast<double>(n);
+        if (n == 0) return 0;
+        return total_active_time_ns_.load() / n;
     }
 
     // Reset (single-threaded only)
